Fixed parse_str abstraction location to ./parser/mod.rs and updated inline documentation
